20230131_Qtapp/requirements.txt
PyJWT>=2.6.0
PyQt5>=5.15.7
